# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sandboxd /v1 API
  version: "0.3.0"
  description: >
    The public, tenant-scoped API for sandboxd. This is the only surface a
    console or external integration should use — the internal /sandbox and
    /wake endpoints are not part of this contract. In the local Docker profile,
    requests may use a server session cookie, a bearer API token, or (when
    explicitly disabled) no credential. In the production AKS profile, requests
    use the server session created by OIDC authorization-code PKCE in the
    configured Entra tenant; local passwords and API keys are not production
    authentication mechanisms. Production ownership is keyed to the immutable
    Entra OID and roles are sandboxd.user or sandboxd.admin.
servers:
  - url: http://127.0.0.1:9090
    description: Default loopback bind (docker-compose maps SANDBOXD_API_BIND here)
security:
  - sessionCookieAuth: []
  - bearerAuth: []
  - {} # local auth-disabled mode only
...
paths:
...
  /v1/git-credentials:
    post:
      operationId: createGitCredential
      tags: [apps]
      summary: Store an owner-scoped Git credential (token write-only, encrypted at rest). Returns metadata only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, token]
              properties:
                name: { type: string }
                host: { type: string, description: e.g. github.com }
                username: { type: string }
                token: { type: string, description: PAT — write-only, never returned }
      responses:
        "201": { description: Created (metadata only; token never returned) }
        "400": { description: Invalid request }
        "409": { description: A credential with that name already exists }
        "503": { description: Credential store not configured }
    get:
      operationId: listGitCredentials
      tags: [apps]
      summary: List the tenant's Git credentials (metadata only — no token)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  credentials:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        host: { type: string }
                        username: { type: string }
                        token_set: { type: boolean }
                        created_at: { type: string }
                        updated_at: { type: string }
  /v1/git-credentials/{id}:
    delete:
      operationId: deleteGitCredential
      tags: [apps]
      summary: Delete a Git credential (owner-scoped)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "404": { description: No such credential }
...
  /v1/sandboxes/{id}/files:
    parameters: [ { $ref: "#/components/parameters/Id" } ]
    get:
      operationId: listSandboxFiles
      tags: [sandboxes]
      summary: List workspace files
      parameters:
        - { in: query, name: path, schema: { type: string }, description: "Directory relative to the app dir (default '.')." }
        - { in: query, name: recursive, schema: { type: boolean } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  path: { type: string }
                  recursive: { type: boolean }
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        path: { type: string, description: "Relative to the app dir." }
                        type: { type: string, enum: [file, dir] }
                        size: { type: integer }
    put:
      operationId: putSandboxFile
      tags: [sandboxes]
      summary: Write a workspace file (atomic)
      description: >
        Writes one file into the sandbox app directory (`workspace/app` inside the
        runtime). The `path` query parameter is app-relative for every runtime, so
        `path=sandbox.yaml` writes the manifest runtimed reads. Reserved subtrees
        (`.runtimed/`, `lost+found/`) are refused. The request body is the RAW file
        content (opaque bytes), not JSON.
      parameters:
        - in: query
          name: path
          required: true
          schema: { type: string }
          description: >
            File path relative to the app directory (e.g. `sandbox.yaml` or
            `src/main.ts`). Absolute paths and `..` are rejected.
      requestBody:
        required: true
        description: Raw file content (opaque bytes) — not JSON.
        content:
          text/plain:
            schema: { type: string }
      responses:
        "200": { description: Written }
        "400": { description: Invalid path (absolute, traversal, or reserved subtree) }
        "413": { description: File exceeds the size limit }
...
components:
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: sbx_session
      description: >
        HttpOnly, host-only server session cookie. In production it is issued
        after the OIDC authorization-code PKCE flow; the cookie is not a bearer
        credential for preview hosts.
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
...